Move sockets to utils
