+++ backend/endpoints/scan.py
@@ -6,8 +6,10 @@ from logger.logger import log
 from utils import fs, fastapi
 from utils.exceptions import PlatformsNotFoundException, RomsNotFoundException
 from handler import dbh
-from handler.socket_manager import socket_server
+from utils.socket import socket_server
 from utils.cache import redis_client, redis_url, redis_connectable
+
+scan_queue = Queue(connection=redis_client)
 
 
 async def scan_platforms(paltforms: str, complete_rescan: bool):
@@ -80,9 +82,6 @@ async def scan_handler(_sid: str, platforms: str, complete_rescan: bool = True):
 
     # Run in worker if redis is available
     if redis_connectable:
-        scan_queue = Queue(connection=redis_client)
-
-        # Queue up a job for all platform
         scan_queue.enqueue(scan_platforms, platforms, complete_rescan)
     else:
         await scan_platforms(platforms, complete_rescan)
