+++ backend/handler/igdb_handler.py
@@ -250,8 +250,8 @@ class TwitchAuth:
         token = cache.get("twitch_token")
         token_expires_at = cache.get("twitch_token_expires_at")
 
-        if not token or time.time() > token_expires_at:
+        if not token or time.time() > float(token_expires_at):
             log.warning("Twitch token invalid: fetching a new one...")
             return self._update_twitch_token()
 
-        return self.oauth_token
+        return token
